Fixing problem in 6.0-backup pushbuild.
Test rpl_ndb_ctype_ucs2_def requires non-standard settings for default character 
set and collation in the test database. This is achieved by specifying these 
defaults as server options in the rpl_ndb_ctype_ucs2_def-master.opt file.

However, if in some test which runs before rpl_ndb_ctype_ucs2_def the test 
database is dropped, and created (as is the case when RESTORE is executed) then 
the default character set and collation are stored in db.opt file inside test/ 
directory. In that case, even if the server is restarted, defaults will be read 
from the db.opt file, not from server options. This makes the test to fail.

This is solved by dropping and re-creating test database inside 
rpl_ndb_ctype_ucs2_def test. This way, when the database is created, the 
character set and collation are taken from server options as they should.

This issue is similar to that reported in BUG#34346 but the fix is different 
with the advantage of less changes made to the test case.