Bug#34381 Client application cannot connect to server on Windows Vista The problem was that the server is creating an IPv6 socket and listens only to IPv6 traffic. If an older client application tries to connect to server using IPv4 address (like 127.0.0.1 for localhost), it fails.
The solution is to clear IPV6_V6ONLY socket option before bind() - in 
this case we will have dual-mode socket that accepts both IPv4 and IPv6
traffic. 

This is done for all operating systems where IPV6_V6ONLY is defined.
Although RFC3493 recommends to set IPV6_V6ONLY "off" by default and many 
follow the recommendation, on Vista it is "on" for compatibility reasons 
(separate IPv4 and IPv6 stacks in prior version). On FreeBSD  the setting
can be controlled via sysctrl