BUG#34346 (ctype_latin1_de test fails if test database is recreated):
This patch reverts the changes introduces by previous one since a different
solution of the problem was implemented. Now all tests which might re-create
the test database (e.g., the backup tests) take care of cleaning the test/ dir
so that it doesn't contain the db.opt file. Thus this test can work in the
form in which it was originaly designed.