Bug#17954 Threads_connected > Threads_created
The problem is that insert delayed threads are counted as connected
but not as created, leading to a Threads_connected value greater then
the Threads_created value.

The solution is to enforce the documented behavior that the 
Threads_connected value shall be the number of currently
open connections and that Threads_created shall be the
number of threads created to handle connections.