Bug#24289 Status Variable "Questions" gets wrong values with Stored Routines
When running Stored Routines the Status Variable "Questions" was wrongly incremented.
According to the manual it should contain the "number of statements that clients have sent
to the server"

Introduced a new status variable 'questions' to replace the query_id
variable which currently corresponds badly with the number of statements
sent by the client.

The new behavior is ment to be backward compatible with 4.0 and at the
same time work with new features in a similar way.